+++ tests/testthat/test-qenv-class.R
@@ -1,13 +1,13 @@
 testthat::describe("methods::new(qenv)", {
   testthat::it("creates a locked environment", {
-    expect_true(environmentIsLocked(teal.code::get_env(methods::new("qenv"))))
+    expect_true(environmentIsLocked(as.environment(methods::new("qenv"))))
   })
 
   testthat::it("creates a locked environment when .xData is manually defined", {
     new_env <- new.env()
     expect_false(environmentIsLocked(new_env))
 
-    expect_true(environmentIsLocked(teal.code::get_env(methods::new("qenv", .xData = new_env))))
+    expect_true(environmentIsLocked(as.environment(methods::new("qenv", .xData = new_env))))
   })
 
   testthat::it("throws error when id and code length doesn't match", {